Rework sub-parser default application:   -- for sub_argument action apply defaults for _all_ subparsers      whether sub-parsers 'used' during parsing or not.   -- for sub_command action only apply defaults for sub_command(s)      that are parsed.
Reasoning: sub_command actions usually apply to a positional agument
after which the whole remaining arguments are parsed using that single
parser so only the defaults for this sub-command would seem relevant.

sub_argument actions apply to options which can provide arguments
piecemeal to various sub-parsers. Hence defaults for the whole set of
sub-parsers is more likely to be relevant.